Move generic part of MALPM.Create module to LPM module
