+++ src/malpm/Create.jl
@@ -1,9 +1,9 @@
 module Create
 
-
-using Utilities: Gender, unknown, female, male
-using XAgents: Town, PersonHouse, Person, undefinedHouse, setAsPartners! 
+using XAgents: Town, PersonHouse, Person 
 using MultiAgents: ABM, attach2DData!
+
+using LPM.Demography.Create: createUKTowns, createUKPopulation
 
 import SomeUtil: AbstractExample
 
@@ -20,54 +20,6 @@ struct LPMUKDemography <: Demography end
 
 "This is an attemp for improved algorthimic translation"
 struct LPMUKDemographyOpt <: Demography end 
-
-### 
-
-function createUKTowns(pars) 
-
-    uktowns = Town[] 
-    
-    for y in 1:pars.mapGridYDimension
-        for x in 1:pars.mapGridXDimension 
-            town = Town((x,y),density=pars.ukMap[y,x])
-            push!(uktowns,town)
-        end
-    end
-
-    uktowns
-end
-
-function createUKPopulation(pars) 
-
-    population = Person[] 
-
-    for i in 1 : pars.initialPop
-        ageMale = rand((pars.minStartAge:pars.maxStartAge))
-        ageFemale = ageMale - rand((-2:5))
-        ageFemale = ageFemale < 24 ? 24 : ageFemale
-        
-        rageMale = ageMale + rand(0:11) // 12     
-        rageFemale = ageFemale + rand(0:11) // 12 
-
-        # From the old code: 
-        #    the following is direct translation but it does not ok 
-        #    birthYear = properties[:startYear] - rand((properties[:minStartAge]:properties[:maxStartAge]))
-        #    why not 
-        #    birthYear = properties[:startYear]  - ageMale/Female 
-        #    birthMonth = rand((1:12))
-
-        newMan = Person(undefinedHouse,rageMale,gender=male)
-        newWoman = Person(undefinedHouse,rageFemale,gender=female)   
-        setAsPartners!(newMan,newWoman) 
-        
-        push!(population,newMan);  push!(population,newWoman) 
-
-    end # for 
-
-    population
-
-end # createUKPopulation 
-
 
 "create UK demography"
 function createUKDemography(properties) 
@@ -98,8 +50,6 @@ function createUKDemography(properties)
 
     [ukTowns,ukHouses,ukPopulation]
 end 
-
-
 
 
 end # Create
